feat: fetchWithAuth util 함수 구현
