+++ src/stores/cartItems.ts
@@ -1,5 +1,5 @@
 import { selector } from "recoil";
-import { getCartItems } from "../apis";
+import { getCartItems } from "../apis/cart";
 import { CartItemType } from "../types";
 
 export const cartItemsState = selector<CartItemType[]>({
